Minor Tweaks
Current score board to show score as you progress in the game.
Diamond pickup sound effect and particle effect tweaks.
Platform blocks are only generated if one is destroyed. This is to avoid
the game from running out of platforms, or overloaded with platforms.
Score is increased by one if you progress through one platform block, or
pick up a diamond.